Change return type of `SigningKey::as_bytes()`
