+++ src/jwk.rs
@@ -36,11 +36,13 @@
 
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use sha2::digest::{Digest, Output};
-use zeroize::Zeroize;
 
-use core::{fmt, ops};
+use core::fmt;
 
-use crate::alloc::{Cow, String, ToString, Vec};
+use crate::{
+    alg::SecretBytes,
+    alloc::{Cow, String, ToString, Vec},
+};
 
 /// Type of a [`JsonWebKey`].
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
@@ -170,70 +172,6 @@ impl JwkError {
     }
 }
 
-/// Generic container for secret bytes, which can be either owned or borrowed.
-/// If owned, bytes are zeroized on drop.
-///
-/// Comparisons on `SecretBytes` are constant-time, but other operations (e.g., deserialization)
-/// may be var-time.
-///
-/// # Serialization
-///
-/// Represented in human-readable formats (JSON, TOML, YAML, etc.) as a base64-url encoded string
-/// with no padding. For other formats (e.g., CBOR), `SecretBytes` will be serialized directly
-/// as a byte sequence.
-#[derive(Clone)]
-pub struct SecretBytes<'a>(Cow<'a, [u8]>);
-
-impl<'a> SecretBytes<'a> {
-    /// Creates secret bytes from a borrowed slice.
-    pub fn borrowed(bytes: &'a [u8]) -> Self {
-        Self(Cow::Borrowed(bytes))
-    }
-
-    /// Creates secret bytes from an owned `Vec`.
-    pub fn owned(bytes: Vec<u8>) -> Self {
-        Self(Cow::Owned(bytes))
-    }
-}
-
-impl fmt::Debug for SecretBytes<'_> {
-    fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
-        formatter
-            .debug_struct("SecretBytes")
-            .field("len", &self.0.len())
-            .finish()
-    }
-}
-
-impl Drop for SecretBytes<'_> {
-    fn drop(&mut self) {
-        // if bytes are borrowed, we don't need to perform any special cleaning.
-        if let Cow::Owned(bytes) = &mut self.0 {
-            Zeroize::zeroize(bytes);
-        }
-    }
-}
-
-impl ops::Deref for SecretBytes<'_> {
-    type Target = [u8];
-
-    fn deref(&self) -> &Self::Target {
-        &*self.0
-    }
-}
-
-impl AsRef<[u8]> for SecretBytes<'_> {
-    fn as_ref(&self) -> &[u8] {
-        &*self
-    }
-}
-
-impl PartialEq for SecretBytes<'_> {
-    fn eq(&self, other: &Self) -> bool {
-        subtle::ConstantTimeEq::ct_eq(self.as_ref(), other.as_ref()).into()
-    }
-}
-
 impl Serialize for SecretBytes<'_> {
     fn serialize<S: Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
         base64url::serialize(self.as_ref(), serializer)
@@ -242,7 +180,7 @@ impl Serialize for SecretBytes<'_> {
 
 impl<'de> Deserialize<'de> for SecretBytes<'_> {
     fn deserialize<D: Deserializer<'de>>(deserializer: D) -> Result<Self, D::Error> {
-        base64url::deserialize(deserializer).map(SecretBytes)
+        base64url::deserialize(deserializer).map(SecretBytes::new)
     }
 }
